requirements.txt
asgiref==3.8.1
Django==4.2.10
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==24.2
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.1
whitenoise==6.9.0
